full_dj/pro/requirement.txt
Django==3.0.2
django-restframework==0.0.1
mysql==0.0.2
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0